pulumi_code/pyproject.toml
[tool.poetry]
name = "pulumi_code"
version = "0.1.0"
description = ""
authors = ["takat0-h0rikosh1 <takato.h0rikosh1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pulumi = "^3.81.0"
pulumi-aws = "^6.0.4"
pulumi-docker = "^4.4.0"
PyYAML = "^6.0.1"
boto3 = "^1.28.44"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
